+++ mp3_metadata.py
@@ -6,7 +6,7 @@ import re
 import sys
 from os.path import basename, dirname, isfile, join
 from pathlib import Path
-from typing import Any, List, Optional, Tuple
+from typing import List, Optional, Tuple
 
 import music_tag
 from colorama import Back, Fore
@@ -15,40 +15,13 @@ from dateutil.parser import ParserError
 
 from config import IMG_DIR, IS_DEBUG
 from music_api import ReleaseRecording, download_album_artwork, get_track_info
+from user_interaction import choose_recording, get_user_input, print_suggestions
 
 logger = logging.getLogger("XP3")
 logger.setLevel(logging.DEBUG if IS_DEBUG else logging.INFO)
 
 PATTERN_ILLEGAL_CHARS = r'[\\/:*?"<>|]'
 STRINGS_TO_REMOVE = ["with Lyrics", "Lyrics", "720p", "1080p", "Video", "LYRICS"]
-
-
-def get_user_input(prompt: str, default: Any) -> str:
-    """Handles user input for functions that require interactivity
-
-    Args:
-        prompt (str): The message who'll be shown to the user
-        default (Any): The default value to be suggested to the user
-
-    Returns:
-        str: The chosen user input
-    """
-    if prompt.endswith(":"):
-        prompt = prompt[:-1]
-    prompt = prompt.strip()
-    if default:
-        if str(default).lower() == "y":
-            prompt = f"{prompt} [Y/n]"
-        elif str(default).lower() == "n":
-            prompt = f"{prompt} [y/N]"
-        else:
-            prompt = f"{prompt} [{default}]"
-    user_input = input(prompt + ": ")
-
-    if not user_input and default:
-        return default
-
-    return user_input
 
 
 def extract_date_from_string(string: str) -> Optional[datetime.datetime]:
@@ -173,35 +146,6 @@ def convert_from_filename(filename: str) -> str:
     return filename
 
 
-# TODO - consider changing this with `pick`
-def print_suggestions(
-    recordings: List[ReleaseRecording],
-    artist: str,
-    title: str,
-    suggested_recording: int,
-):
-    """Prints the suggestions recordings.
-
-    Args:
-        recordings List[str, int, int]: The suggested recordings to print
-        artist (str): The artist of the track
-        title (str): The title of the track
-        suggested_recording (int): The default recording to choose.
-    """
-    print("--------------------")
-    print(f"Choose the correct album for {artist} - {title}:")
-    print(" -1: Skip album metadata")
-    print("--------------------")
-    print(" 0 : Type metadata manually")
-    print("--------------------")
-    for index, recording in enumerate(recordings):
-        if suggested_recording == index:  # Highlight suggested album
-            print(Fore.BLUE, Back.WHITE, end="")
-        print(f"{index + 1} : {recording.album}")
-        print(f" >> year : {recording.year}, track : {recording.track}" + Fore.RESET + Back.RESET)
-        print("--------------------")
-
-
 def get_album_artwork_path(band: str, song: str, album: str = "") -> Tuple[str, str]:
     """Gets path for the album artwork image.
     If it exists, the album will be used. Otherwise, the song will be used.
@@ -212,32 +156,6 @@ def get_album_artwork_path(band: str, song: str, album: str = "") -> Tuple[str, 
     name_for_art = album if album else song
     album_artwork_path = join(IMG_DIR, f"{band} - {name_for_art}.png")
     return album_artwork_path, name_for_art
-
-
-def choose_recording(recordings: List[ReleaseRecording], suggested_recording: int) -> Optional[ReleaseRecording]:
-    """Chooses a recording interactivly using user input.
-
-    Args:
-        recordings (List[ReleaseRecording]): List of suggested recordings.
-        suggested_recording (int): Default index for recording from the recordings list.
-
-    Returns:
-        ReleaseRecording: Chosen Recording
-    """
-    recording_index = get_user_input("Enter the correct album number", default=suggested_recording + 1)
-    recording_index = int(recording_index)
-    # TODO - validate recording_index
-    if not recordings:
-        return None
-    if recording_index == -1:
-        return None  # No album information needed
-    if recording_index == 0:
-        # TODO - validate input
-        album = get_user_input("Enter album name", default=recordings[suggested_recording].album)
-        year = int(get_user_input("Enter album year", default=recordings[suggested_recording].year))
-        track = int(get_user_input("Enter album track", default=recordings[suggested_recording].track))
-        return ReleaseRecording(album, year, "", track, "")
-    return recordings[recording_index - 1]
 
 
 def get_title_from_path(file_path: str) -> str:
@@ -482,16 +400,26 @@ class MP3MetaData:
         """Get title (band - song)"""
         return self.band + " - " + self.song
 
+    def update_fields_from_recording(self, recording: Optional[ReleaseRecording] = None):
+        """
+        If recording is given, updates year, album, track from it
+        """
+        if not recording:
+            return
+        self.year = recording.year
+        self.album = recording.album
+        self.track = recording.track
+
     def update_missing_fields(self, interactive: bool = False, keep_current_metadata: bool = False):
         """Updates missing mp3 metadata fields.
         Args:
             interactive (bool, optional): Should run in interactive mode, get user feedback regarding title fixes.
                                           Defaults to False.
             keep_current_metadata (bool, optional): States whether to use existing metadata.
-                                                    USED ONLY IF interactive=False!
                                                     Defaults to False.
         """
         if not self.title:
+            logger.debug("No title. Returning...")
             return
 
         # Check if there're missing fields
@@ -500,9 +428,9 @@ class MP3MetaData:
             if not interactive and keep_current_metadata:
                 logger.debug("Keeping current metadata for %s", self.title)
                 return
-            if not interactive:
+            if not interactive or keep_current_metadata:
                 should_use_existing_metadata = keep_current_metadata
-            else:  # Interactive mode
+            else:  # Interactive mode, keep_current_metadata is False
                 should_use_existing_metadata = get_user_input(
                     prompt=f"""Metadata already set.
 {self.title}
@@ -529,26 +457,15 @@ Skip?""",
 
         if not interactive:
             if not recordings:
-                self.album = ""
-                self.year = 0
-                self.track = 0
                 return
             suggested_album = max(suggested_album, 0)
-            self.album = recordings[suggested_album].album
-            self.year = recordings[suggested_album].year
-            self.track = recordings[suggested_album].track
+            self.update_fields_from_recording(recordings[suggested_album])
             return
 
         print_suggestions(recordings, artist, title, suggested_album)
         chosen_recording = choose_recording(recordings, suggested_album)
-        if chosen_recording is None:
-            self.album = ""
-            self.year = 0
-            self.track = 0
-        else:
-            self.album = chosen_recording.album
-            self.year = chosen_recording.year
-            self.track = chosen_recording.track
+        if chosen_recording is not None:
+            self.update_fields_from_recording(chosen_recording)
 
         logger.debug("Album: %s, year: %d, track: %d", self.album, self.year, self.track)
